Update patch set 1
Patch Set 1:

Looks like a policy change from 10.0 resets voice-call volume on call hangup to 0.4.  So this is only an issue if the volume is left at 0.4 at the end of the call, since other volumes will force the volume to be reset.

In any event, this covers all cases.

Patch-set: 1