Update patch set 1
Patch Set 1:

ask arcee why he merged it in the other one maybe the resampler issue is still a problem when it tries to access the device. device init> directly to the audio socket stack but lpa put it in the a stream of  magic.  then goes to a2dp stack thus it takes a lot longer to start up . and causes an error on the initial  hopefully the 4.2.2 will fix it

Patch-set: 1