Update patch set 6
Patch Set 6:

I tested that out.  Doesn't work as-is for dock audio...

Not sure if that's really a reason not to submit this patch - this patch brings us back to cm9/10 level functionality.  Enhancing it to output to both speaker and dock redirection simultaneously for ringtones is certainly a nice move, but also not strictly necessary?

Patch-set: 6