Update patch set 1
Patch Set 1:

I'm able to fairly reliably reproduce the jf calls with no audio (in either direction) issue.

With this hack, I was able to make 30x test calls without the problem showing (normally it shows within 5-10 calls).

More testing needed to verify if this is the right direction.

Patch-set: 1