Update patch set 3
Patch Set 3:

Oh this is marvellous - the presence of the additional debug has tipped things so that the bug appears almost all the time (I flipped the #if 1 to #if 0 so I would know when a call was good/bad).  I wouldn't believe it myself if I hadn't just flapped the #if between 0 and 1 and few times (rebooting and retesting each time) to check I wasn't losing it.  Remove the extra debug and behaviour is the same as pre-hack..

Race city, it would seem.  This might help explain why some users the issue and some not at all.

Anyway, debug outputs...

Patch-set: 3