alsa_sound: Maintain separate mute states for each call type.
Mute is not proapagated to CS call session during
SRVCC handover.

This is due to mute state is not updated to CS call
session when SRVCC handover happens.

Fix is to maintain different mute states to different
call types, so that mute state will be updated on
each call type independently.

CRs-Fixed: 516986

Change-Id: I88a589ce74bba3e4018e0455f0008b5f1cb002e8