alsa_sound: Clear TX cal for FM use case
For FM usecases RX and TX paths are started.
For TX path there is no device associated with it.
When TX path is started the driver queries for the
TX_CAL and it was getting the calibration data of
previous TX device.
Since FM TX path is not associated with any device,
change clears out the TX_CAL before starting FM usecase.

Change-Id: Id134f790088cbf66c1f78b1b557ae8bf770ae643
CRs-fixed: 520762