Update patch set 1
Patch Set 1: Code-Review+1

as the commit message says. Without this small change no apps can access the microphone. Only phone calls work. The bug appeared the same day this was merged back in 10.1. I reverted it back then not knowing if it was the correct way to ix the issue. But since no one else has seemed to address this any other way then using this commit I figured I would submit it for review. I don't really know what the point of the original commit was. And there are no regressions on any devices I work on by having this in place...

Patch-set: 1
Label: Code-Review=+1