Update patch set 1
Patch Set 1:

A few things...

BOARD_HAVE_DOCK_USBAUDIO is removed in this patch, but both jf-common and d2-common still have it defined. No harm with that though, as it doesn't do anything anyway now.

Also, there's nothing in Android.mk that defines SAMSUNG_AUDIO, which is why it was removed from msm8960-common here: http://review.cyanogenmod.org/#/c/47852/

Coincidentally, Steve submitted this patch the same day that was removed, and it adds SAMSUNG_AUDIO. However, a define for it in Android.mk is still missing.

So a BOARD_HAVE_SAMSUNG_AUDIO check should be added here to Android.mk, and a BOARD_HAVE_SAMSUNG_AUDIO := true should be added in msm8960-common.

I'd be surprised if this actually worked for Steve as it is. :)

Patch-set: 1
Reviewer: Arne Coucheron <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0