Update patch set 1
Patch Set 1:

I've nothing to say about this patch, but is there an alternative? My U8800 features FM radio functionality, but it was never implemented in stock ROMs. I've tried getting it to work with STE fmradio in cm-10.1, but I only got as far as the radio searching. No audio was outputted.

Patch-set: 1