Update patch set 1
Patch Set 1: Code-Review+1

Im sorry for bumping this up, 

cdesai repiled in the xperia z thread : The current kernel isn't outdated at all, it has been updated to a recent upstream (codeaurora) branch, which means, it is close to what an OEM would base their kernels on.
However, the CM kernel isn't based off sony's 4.3, as it dropped after we had updated it.

We are working on updating the blobs, however, there is an issue, calling is completely broken with the new blobs.
3G/Texts are fine, but calls aren't.
Haven't got a chance to look at it again yet, as we're all quite busy (personally speaking, I barely have time these days, and maintaining multiple devices makes it even less) - can't even imagine what it would be like for Kali-/FXP

I Hope that clears any doubts you guys had.

http://forum.xda-developers.com/showpost.php?p=50487710&postcount=314

while this does say pretty much what's wrong with it 

cant we push this to temp fix the audio-stuttering ?

It is going to be close to 3 months...

Patch-set: 1