Update patch set 1
Patch Set 1:

I can tell you with 100% certainty that this is the wrong fix. You're increasing the buffer size of a stream to compensate for jitter elsewhere in the system. You need to fix the source of the latency. We've dealt with issues like this on nearly every device and yes, it's hard, but please do it right.

Patch-set: 1