Update patch set 2
Patch Set 2:

The new patchset is a port of the original changes from the manufacturer from https://github.com/mitwo-dev/platform_hardware_qcom_audio

I dunno if we can merge it as is.
It won't break anything but the code may be a little bit messy. It would be great if someone with much audio-caf experience could take a look.

Patch-set: 2