Update patch set 4
Patch Set 4:

after taking a close look at the code of libaudioamp I don't see how this is helpful.

None of the patches could be moved to this lib because the api-calls are way too late or there just aren't any where changes are needed.

And if you look at the code for es325 codec you'll see that it basically does very similar things.

That said I don't see a way of exporting these patches without a total rewrite of audio_alsa to export all codecs into external libraries.

Patch-set: 4