Update patch set 4
Patch Set 4:

yes I checked the whole code.
look at the function "amplifier_set_devices" for an example.

It gets called after the device-choose code of audio-caf.
But as you can see in this patchset all the modifications we need to make needs to be done before at many different locations.

Patch-set: 4