Update patch set 4
Patch Set 4:

ok it would be useful to export the big functions to libaudioamp. since the header is in the devicetree I can easily export new functions.

This way the big code would be exported, ok.
But the calls themself wouldn't change that much.

I'll upload a new patchset soon.

Patch-set: 4