Update patch set 9
Patch Set 9:

@David Hacker: The quality of the code is my main issue with this patchset, not that it works. I'd prefer not to merge in a few hundred lines of (1) device-specific code. Especially if most of that code is doing config in code, rather than the config file.

@Michael: I'm not sure how to explain it further to you. My comments from previously still stand. You should extend the libaudioamp API with a doRouting function such that the calls to the "doRouting_Audience_Codec" function can be moved into it.
Additionally, the extra custom devices you are defining appear to be unnecessary. You *should* be able to use existing snd_soc_msm devices and use cases to make it work. The config system is flexible enough such that it IS possible. I'm not sure what your current issue with actually *doing* it is, but I assure you that is possible to do without crashing.

If this is for Xiaomi Mi2 support, I think we might have one in the office. If we do, I can try rewriting this patchset if I find some free time.

Patch-set: 9